requirements.txt
numpy==1.26.4
opencv-python-headless==4.8.1.78
Pillow>=10.0.0
PyYAML>=6.0
tqdm>=4.65
matplotlib>=3.7
tensorboard>=2.14
pandas>=2.0
scikit-image>=0.21
pypdf>=4.0
